Remove functionality for defining airlines per user
This was a bad design decision from the beginning and led to a lot of
additional code that needed to pass on the current user deep into the
model and services.

The new approach is to use the Openflights airline list as base and then
have an (internal) way to overwrite the airlines in that list with
"better" airlines that match the reality. This allows us to keep the
vast amount of data provided by the Openflights list while still being
able to add airlines that Openflights doesn't know about (or has gotten
wrong)